pyproject.toml
[project]
name = "domain-qa-chatbot"
version = "0.1.0"
description = "IEOR 4576 Project 1: Domain Q&A Chatbot (FastAPI + eval harness)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "litellm>=1.44.0",
  "httpx>=0.27.0",
  "google-auth>=2.30.0",
  "google-cloud-aiplatform>=1.63.0",
]

[project.scripts]
serve = "app.main:cli_serve"
eval = "eval.run_eval:main"

[tool.uv]
# uv will generate uv.lock after you run `uv run ...` or `uv sync`
